requirements.txt
numpy==1.23.5
opencv-python==4.7.0.72
pandas==1.5.3
scikit-image==0.19.3
torch==2.0.1
torchvision==0.15.2
Pillow==9.4.0
